pyproject.toml
[tool.poetry]
name = "cosupport-rag"
version = "0.1.0"
description = ""
authors = ["Davit98 <dvmartirosyan@gmail.com>"]
readme = "README.md"
packages = [{include = "cosupport_rag"}]

[tool.poetry.dependencies]
python = "3.10.*"
langchain-community = "0.2.4"
faiss-cpu = "1.8.0"
langchain-openai = "0.1.8"
pydantic = "2.7.3"
pydantic-settings = "2.3.2"
langchainhub = "0.1.18"
langchain-pinecone = "0.1.1"
streamlit = "1.35.0"
pillow = "10.3.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
notebook = "7.0.2"
ipywidgets = "8.1.0"
widgetsnbextension = "4.0.8"
pandas = "2.2.2"
matplotlib = "3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
